setup.py/packaging improvements
On windows include the quantlib dlls (and optionally the c runtime dlls, which
can be useful if not using the default version of visual studio) in the quantlib
package. This means that it can be installed from an egg or wheel without
having to also have the quantlib dll somewhere else on the system path.

Also some minor changes to reduce the amount of manual changes
required to setup.py and generate_symbols.py for swiching between x86
and x64.